// restaurant/src/app/generated-api/model/table.ts
/**
 * Swagger Restaurant
 * This is a sample Restaruant
 *
 * The version of the OpenAPI document: 1.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Table { 
    number: number;
    capacity: number;
    status?: Table.StatusEnum;
}
export namespace Table {
    export type StatusEnum = 'booked' | 'free';
    export const StatusEnum = {
        Booked: 'booked' as StatusEnum,
        Free: 'free' as StatusEnum
    };
}